api: Change policy id to uint32
Address long standing TODO's by changing remote_policies to uint32 to
avoid type conversions on grpc clients. Also remove the uniqueness
validation to save on computation cost as duplicate entries are harmless
(we insert remote policy IDs to a hash map which removes duplicates).

+++ cilium/network_policy.h
@@ -30,7 +30,7 @@ public:
 
   virtual bool useProxylib(std::string& l7_proto) const PURE;
 
-  virtual bool Matches(absl::string_view sni, uint64_t remote_id) const PURE;
+  virtual bool Matches(absl::string_view sni, uint32_t remote_id) const PURE;
 
   virtual bool allowed(const envoy::config::core::v3::Metadata& metadata) const PURE;
 
@@ -54,16 +54,16 @@ class PolicyInstance {
 public:
   virtual ~PolicyInstance() = default;
 
-  virtual bool Allowed(bool ingress, uint16_t port, uint64_t remote_id,
+  virtual bool Allowed(bool ingress, uint16_t port, uint32_t remote_id,
                        Envoy::Http::RequestHeaderMap& headers,
                        Cilium::AccessLog::Entry& log_entry) const PURE;
 
   virtual const PortPolicyConstSharedPtr findPortPolicy(bool ingress, uint16_t port,
-                                                        uint64_t remote_id) const PURE;
+                                                        uint32_t remote_id) const PURE;
 
   // Returns true if the policy specifies l7 protocol for the connection, and
   // returns the l7 protocol string in 'l7_proto'
-  virtual bool useProxylib(bool ingress, uint16_t port, uint64_t remote_id,
+  virtual bool useProxylib(bool ingress, uint16_t port, uint32_t remote_id,
                            std::string& l7_proto) const PURE;
 
   virtual const std::string& conntrackName() const PURE;
